Fix the threading thing
The item and shield sync packets relied on the damage type sync packet, but the damage types were only set after the damage type packet is handled, which doesn't happen until all of the packets are decoded for main thread handlers. To fix this, we use the network thread instead for handling the damage type sync and it's fine for the other two to use the main thread.